packages/tasks_loop/pyproject.toml
[project]
name = "tasks-loop"
version = "0.1.0"
description = "Task scheduler for autonomous agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "croniter>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "types-pyyaml>=6.0",
    "types-croniter>=1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tasks_loop"]
